Add x86 assembly for cdef distortion
Around ~7% faster on speed 2. It should be roughly the same for most
speed levels.

It would have been more, but some recent patches seem to have allowed
autovecterization to happen again.
